data_ingestion/requirements.txt
pandas==2.1.1
kafka-python==2.0.2